refactor: extract runPhpStan() to PhpStanTestTrait
